requirements.txt
fastapi==0.104==0.24.0
pydantic==2.5.0
pyodbc==5.1.0
python-dotenv==1.0.0
